+++ src/steps/resource-manager/subscriptions/client.ts
@@ -1,5 +1,5 @@
 import { SubscriptionClient } from '@azure/arm-subscriptions';
-import { Subscription } from '@azure/arm-subscriptions/esm/models';
+import { Subscription, Location } from '@azure/arm-subscriptions/esm/models';
 import {
   Client,
   iterateAllResources,
@@ -20,6 +20,29 @@ export class J1SubscriptionClient extends Client {
       serviceClient,
       resourceEndpoint: serviceClient.subscriptions,
       resourceDescription: 'subscriptions',
+      callback,
+    });
+  }
+
+  public async iterateLocations(
+    subscriptionId: string,
+    callback: (l: Location) => void | Promise<void>,
+  ): Promise<void> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      SubscriptionClient,
+      {
+        passSubscriptionId: false,
+      },
+    );
+    return iterateAllResources({
+      logger: this.logger,
+      serviceClient,
+      resourceEndpoint: {
+        list: async () => {
+          return serviceClient.subscriptions.listLocations(subscriptionId);
+        },
+      },
+      resourceDescription: 'subscriptions.locations',
       callback,
     });
   }