+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Added `azure_location` entities.
 - Added `azure_network_watcher` entities.
 - Added `azure_security_group_flow_logs` entities.
 - Added `azure_sql_server_firewall_rule` entities.